Move useHardwareBackButton to common folder
