+++ ts/features/bonus/bonusVacanze/screens/activation/BonusAlreadyExists.tsx
@@ -7,7 +7,7 @@ import { InfoScreenComponent } from "../../../../../components/infoScreen/InfoSc
 import I18n from "../../../../../i18n";
 import { confirmButtonProps } from "../../components/buttons/ButtonConfigurations";
 import { FooterStackButton } from "../../components/buttons/FooterStackButtons";
-import { useHardwareBackButton } from "../../components/hooks/useHardwareBackButton";
+import { useHardwareBackButton } from "../../../../../hooks/useHardwareBackButton";
 import { bonusVacanzeStyle } from "../../components/Styles";
 import {
   cancelBonusVacanzeRequest,
