+++ ts/features/bonus/bonusVacanze/screens/eligibility/TimeoutEligibilityCheckInfoScreen.tsx
@@ -4,7 +4,7 @@ import { connect } from "react-redux";
 import { Dispatch } from "redux";
 import I18n from "../../../../../i18n";
 import { BaseTimeoutScreen } from "../../components/BaseTimeoutScreen";
-import { useHardwareBackButton } from "../../components/hooks/useHardwareBackButton";
+import { useHardwareBackButton } from "../../../../../hooks/useHardwareBackButton";
 import { cancelBonusVacanzeRequest } from "../../store/actions/bonusVacanze";
 
 type Props = ReturnType<typeof mapDispatchToProps>;
