+++ ts/features/bonus/bonusVacanze/screens/eligibility/isee/BaseIseeErrorComponent.tsx
@@ -10,7 +10,7 @@ import {
   cancelButtonProps,
   confirmButtonProps
 } from "../../../components/buttons/ButtonConfigurations";
-import { useHardwareBackButton } from "../../../components/hooks/useHardwareBackButton";
+import { useHardwareBackButton } from "../../../../../../hooks/useHardwareBackButton";
 import { bonusVacanzeStyle } from "../../../components/Styles";
 
 const inpsDsuHomeUrl =
