+++ ts/features/wallet/onboarding/bancomatPay/screens/add-account/LoadAddBPayComponent.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
 import I18n from "../../../../../../i18n";
-import { useHardwareBackButton } from "../../../../../bonus/bonusVacanze/components/hooks/useHardwareBackButton";
+import { useHardwareBackButton } from "../../../../../../hooks/useHardwareBackButton";
 import { LoadingErrorComponent } from "../../../../../bonus/bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
 
 export type Props = {
