+++ ts/features/wallet/onboarding/privative/screens/choosePrivativeIssuer/ko/PrivativeIssuerKoUnavailable.tsx
@@ -13,7 +13,7 @@ import I18n from "../../../../../../../i18n";
 import { GlobalState } from "../../../../../../../store/reducers/types";
 import { emptyContextualHelp } from "../../../../../../../utils/emptyContextualHelp";
 import { cancelButtonProps } from "../../../../../../bonus/bonusVacanze/components/buttons/ButtonConfigurations";
-import { useHardwareBackButton } from "../../../../../../bonus/bonusVacanze/components/hooks/useHardwareBackButton";
+import { useHardwareBackButton } from "../../../../../../../hooks/useHardwareBackButton";
 import { walletAddPrivativeCancel } from "../../../store/actions";
 
 type Props = ReturnType<typeof mapDispatchToProps> &
