+++ ts/features/wallet/onboarding/satispay/screens/add/LoadAddSatispayComponent.tsx
@@ -4,7 +4,7 @@ import { Dispatch } from "redux";
 import { Satispay } from "../../../../../../../definitions/pagopa/Satispay";
 import I18n from "../../../../../../i18n";
 import { GlobalState } from "../../../../../../store/reducers/types";
-import { useHardwareBackButton } from "../../../../../bonus/bonusVacanze/components/hooks/useHardwareBackButton";
+import { useHardwareBackButton } from "../../../../../../hooks/useHardwareBackButton";
 import { LoadingErrorComponent } from "../../../../../bonus/bonusVacanze/components/loadingErrorScreen/LoadingErrorComponent";
 import {
   addSatispayToWallet,
