+++ ts/utils/useAvoidHardwareBackButton.ts
@@ -1,4 +1,4 @@
-import { useHardwareBackButton } from "../features/bonus/bonusVacanze/components/hooks/useHardwareBackButton";
+import { useHardwareBackButton } from "../hooks/useHardwareBackButton";
 
 export const useAvoidHardwareBackButton = () =>
   useHardwareBackButton(() => true);
